+++ src/Cli/Exception/ConfigException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Soluble\MediaTools\Cli\Exception;
+
+class ConfigException extends \RuntimeException implements ExceptionInterface
+{
+}