+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.3.0 (2019-06-04)
+
+### Added
+
+- Davinci resolve preset
+
+### Upgraded
+
+- Symfony cache, finder and console to 4.3
+
+
 ## 0.2.0 (2019-04-29)
 
 ### Bugfix